server/requirements.production.txt
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
click==6.6
Flask==0.11.1
Flask-Cors==3.0.2
itsdangerous==0.24
Jinja2==2.8
kombu==4.0.2
MarkupSafe==0.23
numpy==1.11.2
protobuf==3.1.0
pytz==2016.10
redis==2.10.5
six==1.10.0
tensorflow-gpu==0.12.0
vine==1.1.3
Werkzeug==0.11.11
